requirements.txt
Flask==1.0.2
gcloud==0.18.3
google-api-core==1.4.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
google-cloud-language==1.0.2
google-cloud-speech==0.36.0
google-cloud-storage==1.12.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.12.1
gunicorn==19.9.0
matplotlib==2.2.3
numpy==1.15.1
oauth2client==4.1.3
oauthlib==2.1.0
pandas==0.23.4
pickleshare==0.7.4
pydub==0.23.0
pytube==9.2.3
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rq==0.12.0
rsa==4.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
SpeechRecognition==3.8.1
urllib3==1.23
